Fix handling of I/O in server
This commit changes the server code to fix a couple errors in how it
handles the streams. First, it changes the loop to return if an empty
bytestring is received from read(), which might happen if the client
closes the connection in a situation where it hasn't issued a directive
to run a command. Second, it closes the writer stream after it's done
handling the command.