Add a command whitelist
This commit adds a whitelist of commands that the server is willing to
run. If a client connects and attempts to run a command not on the list,
the server will send back a response with return code 126 (which is
the same code that shows up in bash when trying to run a non-executable
program or script).