Skip checking exit code after KeyboardInterrupt
This commit fixes a bug where, if the client is keyboard-interrupted,
the exit_code variable isn't defined and the call to sys.exit(exit_code)
fails, producing a NameError and a Python stack trace. The fix is to
call sys.exit() with a fixed value, 130, when the client receives
a KeyboardInterrupt. This is the same code returned by the UNIX utility
sleep when it gets interrupted.