Await process termination in all cases
If the output of a process was being redirected to some destination
stream, the run-remote server would only wait for the process's streams
to reach EOF but would not wait for the process itself to terminate.
This would (rarely) produce a situation where the server would send
a process-end notification (the q command) to the client before
the process had actually terminated, and would do so with a None error
code. Then the client would try to parse that code as an int and
raise an error.

In this commit I've fixed the bug by always awaiting process termination
even if the streams are being redirected.