Add options for logging output and configuration
This commit adds command-line options that can be used to configure
logging. There is a -v/--verbose option which sets the logging level,
as well as an option --log-dest which will set the destination for
logging to stderr or a file, and also an option --log-config which
takes a logging configuration file in the form expected by Python's
logging.config module and can contain arbitrary configuration
directives.