Switch to a general TOML-format configuration file
This commit changes from using a config file for just the logging
system to a general config file that will apply to the whole program.
The configuration file is now a TOML file instead of Python's
configparser format.

The main reason for this change is that the logging module's
documentation says that the fileConfig() method is somewhat deprecated,
or at least may not be updated with future changes, so I wanted to
switch to the more future-proof dictConfig() method. Given that, there's
no reason for the configuration to be in the configparser format. TOML
is a more modern and more precisely defined format, which seems to be
becoming somewhat of a standard for Python projects, so it looked like
a particularly suitable choice.