Add a warning when replacing configured handlers
This commit adds code to emit a warning when the program is run with
logging configuration given in command-line options like --verbose and
--log-dest, if that configuration would override the handlers specified
for the root logger in the configuration file. The warning is emitted
using the handler(s) from the config file which are about to be removed,
so that if someone is examining the output from those handlers, there
will be some indication of why no further logging messages show up
there.