Convert the server and client into classes
This commit introduces a Server class which encapsulates the former
functions of the run_remote.server module, and a Client class which does
the same for the functions of the run_remote.client module. The host
and port are now instance variables instead of being passed directly to
the serve() or run() functions.

This opens the door to having other variables that affect the server
and client which can be stored in instances of their respective classes
and don't need to be passed from function to function.