Dynamically set the sanitizer from the configuration
This commit changes the code that creates the command sanitizer so that
it will dynamically set the sanitizer used based on values in the config
file. This opens up the possibility of using different sanitizers for
different server invocations.

Currently I've only implemented two sanitizers. One is the default,
which is only used if no sanitizer is specified in the config file. It
lets any command that can be found on the PATH run. The other is based
on a whitelist, such that only commands on the whitelist which are also
found on the PATH will be run. Neither sanitizer checks the command's
arguments, but in the future we can also have sanitizers that check
arguments in addition to just the program (first word of the command).

If a sanitizer is specified in the configuration but it can't be
loaded, the server will exit with an error code.